+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Changed
+- AccountList: rename from AccountListDTO
 - IPAddrRange: rename from IPAddrDTO
 - Client.NewRequest: shallow-copy BaseURL to avoid retaining modifications